+++ h2/exceptions.py
@@ -79,9 +79,13 @@ class NoAvailableStreamIDError(ProtocolError):
     pass
 
 
-class NoSuchStreamError(H2Error):
+class NoSuchStreamError(ProtocolError):
     """
     A stream-specific action referenced a stream that does not exist.
+
+    .. versionchanged:: 2.0.0
+       Became a subclass of :class:`ProtocolError
+       <h2.exceptions.ProtocolError>`
     """
     def __init__(self, stream_id):
         #: The stream ID that corresponds to the non-existent stream.